+++ LambdaEngine/Include/Rendering/RenderGraph.h
@@ -175,7 +175,7 @@ namespace LambdaEngine
 			struct
 			{
 				TArray<ResourceBarrierInfo>			BarriersPerSynchronizationStage;
-				TArray<Buffer*>					Buffers;
+				TArray<Buffer*>						Buffers;
 				TArray<uint64>						Offsets;
 				TArray<uint64>						SizesInBytes;
 			} Buffer;
@@ -271,8 +271,8 @@ namespace LambdaEngine
 		CommandList* AcquireComputeCopyCommandList();
 
 		bool GetResourceTextures(const char* pResourceName, Texture* const ** pppTexture, uint32* pTextureView)						const;
-		bool GetResourceTextureViews(const char* pResourceName, TextureView* const ** pppTextureViews, uint32* pTextureViewCount)		const;
-		bool GetResourceBuffers(const char* pResourceName, Buffer* const ** pppBuffers, uint32* pBufferCount)							const;
+		bool GetResourceTextureViews(const char* pResourceName, TextureView* const ** pppTextureViews, uint32* pTextureViewCount)	const;
+		bool GetResourceBuffers(const char* pResourceName, Buffer* const ** pppBuffers, uint32* pBufferCount)						const;
 		bool GetResourceAccelerationStructure(const char* pResourceName, const AccelerationStructure** ppAccelerationStructure)		const;
 
 		virtual void OnWindowResized(TSharedRef<Window> window, uint16 width, uint16 height, EResizeType type) override;
@@ -307,11 +307,11 @@ namespace LambdaEngine
 		void ExecuteRayTracingRenderStage(RenderStage* pRenderStage, PipelineState* pPipelineState, CommandAllocator* pComputeCommandAllocator, CommandList* pComputeCommandList, CommandList** ppExecutionStage);
 
 	private:
-		const GraphicsDevice*				m_pGraphicsDevice;
+		const GraphicsDevice*							m_pGraphicsDevice;
 
 		const Scene*									m_pScene							= nullptr;
 
-		DescriptorHeap*					m_pDescriptorHeap					= nullptr;
+		DescriptorHeap*									m_pDescriptorHeap					= nullptr;
 
 		float32											m_WindowWidth						= 0.0f;
 		float32											m_WindowHeight						= 0.0f;
@@ -322,22 +322,22 @@ namespace LambdaEngine
 		uint32											m_BackBufferCount					= 0;
 		uint32											m_MaxTexturesPerDescriptorSet		= 0;
 
-		Fence*								m_pFence							= nullptr;
-		uint64								m_SignalValue						= 1;
+		Fence*											m_pFence							= nullptr;
+		uint64											m_SignalValue						= 1;
 
 		TArray<ICustomRenderer*>						m_CustomRenderers;
 		TArray<ICustomRenderer*>						m_DebugRenderers;
 
-		CommandAllocator**					m_ppGraphicsCopyCommandAllocators	= nullptr;
-		CommandList**						m_ppGraphicsCopyCommandLists		= nullptr;
-		bool								m_ExecuteGraphicsCopy				= false;
+		CommandAllocator**								m_ppGraphicsCopyCommandAllocators	= nullptr;
+		CommandList**									m_ppGraphicsCopyCommandLists		= nullptr;
+		bool											m_ExecuteGraphicsCopy				= false;
 
-		CommandAllocator**					m_ppComputeCopyCommandAllocators	= nullptr;
-		CommandList**						m_ppComputeCopyCommandLists			= nullptr;
-		bool								m_ExecuteComputeCopy				= false;
+		CommandAllocator**								m_ppComputeCopyCommandAllocators	= nullptr;
+		CommandList**									m_ppComputeCopyCommandLists			= nullptr;
+		bool											m_ExecuteComputeCopy				= false;
 
-		CommandList**						m_ppExecutionStages					= nullptr;
-		uint32								m_ExecutionStageCount				= 0;
+		CommandList**									m_ppExecutionStages					= nullptr;
+		uint32											m_ExecutionStageCount				= 0;
 
 		PipelineStage*									m_pPipelineStages					= nullptr;
 		uint32											m_PipelineStageCount				= 0;
