+++ LambdaEngine/Include/Rendering/RenderGraphEditor.h
@@ -13,138 +13,6 @@
 
 namespace LambdaEngine
 {
-	enum class EEditorPinType : uint8
-	{
-		INPUT					= 0,
-		OUTPUT					= 1,
-		RENDER_STAGE_INPUT		= 2,
-	};
-
-	struct EditorStartedLinkInfo
-	{
-		bool		LinkStarted				= false;
-		int32		LinkStartAttributeID	= -1;
-		bool		LinkStartedOnInputPin	= false;
-		String		LinkStartedOnResource	= "";
-	};
-
-	struct EditorRenderGraphResourceLink
-	{
-		int32		LinkIndex				= 0;
-		int32		SrcAttributeIndex		= 0;
-		int32		DstAttributeIndex		= 0;
-	};
-
-	struct EditorRenderGraphResourceState
-	{
-		String							ResourceName					= "";
-		String							RenderStageName					= "";
-		bool							Removable						= true;
-		ERenderGraphResourceBindingType BindingType						= ERenderGraphResourceBindingType::NONE;
-		int32							InputLinkIndex					= -1;
-		TSet<int32>						OutputLinkIndices;
-	};
-
-	struct EditorResourceStateIdent
-	{
-		String	Name			= "";
-		int32	AttributeIndex	= -1;
-	};
-
-	struct EditorRenderStageDesc
-	{
-		String						Name							= "";
-		int32						NodeIndex						= 0;
-		int32						InputAttributeIndex				= 0;
-		EPipelineStateType			Type							= EPipelineStateType::PIPELINE_STATE_TYPE_NONE;
-		bool						OverrideRecommendedBindingType	= false;
-		bool						CustomRenderer					= false;
-		bool						Enabled							= true;
-		RenderStageParameters		Parameters						= {};
-
-		struct
-		{
-			GraphicsShaderNames			Shaders;
-			ERenderStageDrawType		DrawType;
-			int32						IndexBufferAttributeIndex;
-			int32						IndirectArgsBufferAttributeIndex;
-			bool						DepthTestEnabled;
-		} Graphics;
-
-		struct
-		{
-			String						ShaderName = "";
-		} Compute;
-
-		struct
-		{
-			RayTracingShaderNames		Shaders;
-		} RayTracing;
-
-		uint32							Weight					= 0;
-
-		TArray<EditorResourceStateIdent>		ResourceStateIdents;
-
-		TArray<EditorResourceStateIdent>::Iterator FindResourceStateIdent(const String& name)
-		{
-			for (auto resourceStateIt = ResourceStateIdents.begin(); resourceStateIt != ResourceStateIdents.end(); resourceStateIt++)
-			{
-				if (resourceStateIt->Name == name)
-					return resourceStateIt;
-			}
-
-			return ResourceStateIdents.end();
-		}
-
-		TArray<EditorResourceStateIdent>::ConstIterator FindResourceStateIdent(const String& name) const
-		{
-			for (auto resourceStateIt = ResourceStateIdents.begin(); resourceStateIt != ResourceStateIdents.end(); resourceStateIt++)
-			{
-				if (resourceStateIt->Name == name)
-					return resourceStateIt;
-			}
-
-			return ResourceStateIdents.end();
-		}
-	};
-
-	struct EditorResourceStateGroup
-	{
-		String								Name				= "";
-		int32								InputNodeIndex		= 0;
-		int32								OutputNodeIndex		= 0;
-		TArray<EditorResourceStateIdent>	ResourceStateIdents;
-
-		TArray<EditorResourceStateIdent>::Iterator FindResourceStateIdent(const String& name)
-		{
-			for (auto resourceStateIt = ResourceStateIdents.begin(); resourceStateIt != ResourceStateIdents.end(); resourceStateIt++)
-			{
-				if (resourceStateIt->Name == name)
-					return resourceStateIt;
-			}
-
-			return ResourceStateIdents.end();
-		}
-
-		TArray<EditorResourceStateIdent>::ConstIterator FindResourceStateIdent(const String& name) const
-		{
-			for (auto resourceStateIt = ResourceStateIdents.begin(); resourceStateIt != ResourceStateIdents.end(); resourceStateIt++)
-			{
-				if (resourceStateIt->Name == name)
-					return resourceStateIt;
-			}
-
-			return ResourceStateIdents.end();
-		}
-	};
-
-	struct EditorFinalOutput
-	{
-		String		Name						= "";
-		int32		NodeIndex					= 0;
-		int32		BackBufferAttributeIndex	= 0;
-	};
-
 	class RenderGraph;
 
 	class LAMBDA_API RenderGraphEditor : public EventHandler
@@ -158,8 +26,6 @@ namespace LambdaEngine
 
 		void InitGUI();
 		void RenderGUI();
-
-		RenderGraphStructureDesc CreateRenderGraphStructure(const String& filepath, bool imGuiEnabled);
 
 		virtual void OnButtonReleased(EMouseButton button)						override final;
 		virtual void OnKeyPressed(EKey key, uint32 modifierMask, bool isRepeat) override final;
@@ -199,38 +65,9 @@ namespace LambdaEngine
 
 		void CalculateResourceStateBindingTypes(const EditorRenderStageDesc* pRenderStage, const RenderGraphResourceDesc* pResource, const EditorRenderGraphResourceState* pResourceState, TArray<ERenderGraphResourceBindingType>& bindingTypes, TArray<const char*>& bindingTypeNames);
 
-		bool SaveToFile(const String& renderGraphName);
-		bool LoadFromFile(const String& filepath, bool generateImGuiStage);
-		bool FixLinkForPreviouslyLoadedResourceState(
-			EditorRenderGraphResourceState* pResourceState,
-			int32 attributeIndex,
-			THashTable<int32, EditorRenderGraphResourceState>& loadedResourceStatesByHalfAttributeIndex,
-			THashTable<int32, EditorRenderGraphResourceLink>& loadedResourceStateLinks,
-			TArray<int32>& unfinishedLinksAwaitingStage);
-		void CreateLinkForLoadedResourceState(
-			EditorRenderGraphResourceState* pResourceState,
-			int32 attributeIndex,
-			String& srcStageName,
-			TArray<EditorResourceStateGroup>& loadedResourceStateGroups,
-			THashTable<String, EditorRenderStageDesc>& loadedRenderStagesByName,
-			THashTable<int32, EditorRenderGraphResourceState>& loadedResourceStatesByHalfAttributeIndex,
-			THashTable<int32, EditorRenderGraphResourceLink>& loadedResourceStateLinks,
-			THashTable<String, TArray<int32>>& unfinishedLinks);
+		bool LoadFromFile(const String& renderGraphFileName);
 		void SetInitialNodePositions();
 		void ResetState();
-
-		bool ParseStructure(bool generateImGuiStage);
-		bool RecursivelyWeightParentRenderStages(EditorRenderStageDesc* pChildRenderStage);
-		bool IsRenderStage(const String& name);
-		bool CapturedByImGui(const RenderGraphResourceDesc* pResource);
-		bool FindAndCreateSynchronization(bool generateImGuiStage,
-			const std::multimap<uint32, EditorRenderStageDesc*>::reverse_iterator& currentOrderedRenderStageIt, 
-			const std::multimap<uint32, EditorRenderStageDesc*>& orderedMappedRenderStages, 
-			const EditorRenderGraphResourceState* pCurrentResourceState, 
-			SynchronizationStageDesc* pSynchronizationStage);
-		void CreateParsedRenderStage(RenderStageDesc* pDstRenderStage, const EditorRenderStageDesc* pSrcRenderStage);
-
-		RenderGraphResourceDesc CreateBackBufferResource();
 
 	private:
 		bool												m_GUIInitialized = false;
@@ -256,7 +93,6 @@ namespace LambdaEngine
 		RenderGraphStructureDesc							m_ParsedRenderGraphStructure	= {};
 		bool												m_ParsedGraphDirty				= true;
 		bool												m_ParsedGraphRenderDirty		= true;
-		String												m_ParsingError					= "No Errors";
 
 	private:
 		static int32					s_NextNodeID;
