+++ Sandbox/imgui.ini
@@ -6,7 +6,7 @@ Collapsed=0
 [Window][Render Graph Editor]
 Pos=23,19
 Size=1766,913
-Collapsed=1
+Collapsed=0
 
 [Window][Dear ImGui Demo]
 Pos=96,47
@@ -16,7 +16,7 @@ Collapsed=1
 [Window][Debugging Window]
 Pos=115,90
 Size=642,625
-Collapsed=0
+Collapsed=1
 
 [Window][Graph View]
 Pos=542,83
